Minor changes fixed #65
