+++ libs/planning/planning_py/src/planning/llm/primitive_breakdown.py
@@ -150,58 +150,3 @@ class PrimitiveBreakdown:
 
         return data
     
-
-    def _load_move_transport_preconditions(
-        self, path: str
-    ) -> Tuple[str, Dict[str, str], List[List[str]]]:
-        """
-        Load MOVE_TRANSPORT preconditions from the current YAML schema:
-
-        - llm_predicates: {PREDICATE_KEY: "human readable text", ...}
-        - low_level_primitives: [ { name: MOVE_TRANSPORT, precondition_sets: [...] , ...}, ... ]
-
-        Args:
-            path (str): Path to primitives_preconditions.yaml.
-
-        Returns:
-            Tuple[str, Dict[str, str], List[List[str]]]:
-                description, predicate_text, precondition_sets
-        """
-
-        if yaml is None:
-            raise ImportError("pyyaml is required to read YAML preconditions.")
-        if not os.path.exists(path):
-            raise FileNotFoundError(f"Preconditions file not found: {path}")
-
-        with open(path, "r", encoding="utf-8") as f:
-            cfg = yaml.safe_load(f)
-
-        predicate_text: Dict[str, str] = cfg.get("llm_predicates", {})
-
-        low_level = cfg.get("low_level_primitives", [])
-        mt = None
-        for p in low_level:
-            if isinstance(p, dict) and p.get("name") == "MOVE_TRANSPORT":
-                mt = p
-                break
-
-        if mt is None:
-            raise ValueError("MOVE_TRANSPORT not found in preconditions YAML")
-
-        description: str = mt.get("description", "MOVE_TRANSPORT")
-
-        raw_sets = mt.get("precondition_sets", None)
-        if raw_sets is None:
-            legacy = mt.get("pre-condition", "")
-            raw_sets = [[legacy]] if isinstance(legacy, str) and legacy.strip() else [[]]
-
-        norm_sets: List[List[str]] = []
-        for s in raw_sets:
-            if isinstance(s, list):
-                norm_sets.append(s)
-            elif isinstance(s, str):
-                norm_sets.append([s])
-            else:
-                raise ValueError("Each precondition set must be a list or string.")
-
-        return description, predicate_text, norm_sets
